Implement import script
Implements an import script that will parse and read a rar'd set of CSV files in the provided format. This data is then stored into a SQLite database (as specified by command line parameters).

This also includes starting implementations for the streaming server. The
streaming server will be a combination server that has two main scripts:

1. Streaming data
2. Import script that adds data to the database